[ep-perf] Fix the order of onnx and onnxruntime imports. (#11383)
Fix the order of onnx and onnxruntime imports. Importing onnx before onnxruntime causes a dependency issue in the tensorrt containers that prevents onnxruntime_pybind11_state.so from finding the system libstdc++. This is a workaround to get the EP Perf pipeline working until we can investigate the issue more closely.

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>