Revert "Fix shape-related issues in FuseConv (#12410)" (#13353)
The commit causes subtle perf regressions in image models (caught by
Anubis). Since we are close to the release, reverting this change for
now so that the regression cause analysis doesn't push the release
timeline.

Once the PR is merged, I will re-open the GH issues that the original PR
closed.

### Motivation and Context
Fix regression in ORT 1.13 RC