DML EP enable the provider in the op tests (#13441)
### Description
Enables the DML provider in the op tests to allow for better CI
coverage.



### Motivation and Context
Some of the CI tests for DML were actually running on the CPU because
there was no default DML provider, so it was returning a `nullptr`. This
should add better coverage, and it already uncovered some failures and
asserts hitting in a few tests, which need to be investigated
separately.