Ensure graph resolve occurs after free dimension is overridden (#13634)
### Description
This ensures that the graph is re-resolved after a free dimension shape
is overridden according to session options.

### Motivation and Context
This ensures that shape inference occurs, which is necessary to apply
the optimation and ensure it the session is compatible with bound
shapes. This bug seems to only have affected a small fraction of models.